requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
google-generativeai>=0.3.0
python-dotenv>=0.19.0
pytest>=6.2.4
httpx>=0.23.0